Move args to a seperate file
