+++ gobblin-metrics-libs/gobblin-metrics-base/src/main/java/org/apache/gobblin/metrics/MetricContext.java
@@ -477,7 +477,7 @@ public class MetricContext extends MetricRegistry implements ReportableContext, 
    * @return the {@link ContextAwareHistogram} with the given name
    */
   public ContextAwareHistogram contextAwareHistogram(String name, long windowSize, TimeUnit unit) {
-    ContextAwareMetricFactory.SlidingTimeWindowArgs args = new ContextAwareMetricFactory.SlidingTimeWindowArgs(
+    ContextAwareMetricFactoryArgs.SlidingTimeWindowArgs args = new ContextAwareMetricFactoryArgs.SlidingTimeWindowArgs(
         this.innerMetricContext.getMetricContext().get(), name, windowSize, unit);
     return this.innerMetricContext.getOrCreate(ContextAwareMetricFactory.DEFAULT_CONTEXT_AWARE_HISTOGRAM_FACTORY, args);
   }
@@ -501,7 +501,7 @@ public class MetricContext extends MetricRegistry implements ReportableContext, 
    * @return the {@link ContextAwareTimer} with the given name
    */
   public ContextAwareTimer contextAwareTimer(String name, long windowSize, TimeUnit unit) {
-    ContextAwareMetricFactory.SlidingTimeWindowArgs args = new ContextAwareMetricFactory.SlidingTimeWindowArgs(
+    ContextAwareMetricFactoryArgs.SlidingTimeWindowArgs args = new ContextAwareMetricFactoryArgs.SlidingTimeWindowArgs(
         this.innerMetricContext.getMetricContext().get(), name, windowSize, unit);
     return this.innerMetricContext.getOrCreate(ContextAwareMetricFactory.DEFAULT_CONTEXT_AWARE_TIMER_FACTORY, args);
   }
